+++ src/controllers/nicknames/NicknamesModel.cpp
@@ -8,7 +8,7 @@
 namespace chatterino {
 
 NicknamesModel::NicknamesModel(QObject *parent)
-    : SignalVectorModel<Nickname>(2, parent)
+    : SignalVectorModel<Nickname>(4, parent)
 {
 }
 
@@ -17,7 +17,9 @@ Nickname NicknamesModel::getItemFromRow(std::vector<QStandardItem *> &row,
                                         const Nickname &original)
 {
     return Nickname{row[0]->data(Qt::DisplayRole).toString(),
-                    row[1]->data(Qt::DisplayRole).toString()};
+                    row[1]->data(Qt::DisplayRole).toString(),
+                    row[2]->data(Qt::CheckStateRole).toBool(),
+                    row[3]->data(Qt::CheckStateRole).toBool()};
 }
 
 // turns a row in the model into a vector item
@@ -26,6 +28,8 @@ void NicknamesModel::getRowFromItem(const Nickname &item,
 {
     setStringItem(row[0], item.name());
     setStringItem(row[1], item.replace());
+    setBoolItem(row[2], item.isRegex());
+    setBoolItem(row[3], item.isCaseSensitive());
 }
 
 }  // namespace chatterino